IN-175 fix: layout for Terms and Policy, rename modal into basModal
