+++ src/common/components/baseModal/index.ts
@@ -1,0 +1,1 @@
+export * from './BaseModal'
