+++ src/common/components/index.ts
@@ -1,3 +1,4 @@
+export * from './baseModal'
 export * from './button'
 export * from './calendar'
 export * from './card'
@@ -8,7 +9,6 @@ export * from './header/langSelect'
 export * from './loader'
 export * from './menu'
 export * from './mobileHeader'
-export * from './modal'
 export * from './pagination'
 export * from './popover'
 export * from './radioGroup'
