--- src/common/components/modal/index.ts
@@ -1,1 +1,0 @@
-export * from './Modal'
